add knobs for top-down-analysis metrics
The availability knob actually checks a different register with a
restrictive bitmask reducing the result to a single bit.

The performance metrics knob is only applied to alderlake, even though
it is available since the 10th generation. It is only available for
P-Cores, but currently requested also for E-Cores. During tests this did
not prove to be a problem.