fix x86_adapt_lookup_ci_name max length of strncmp
x86_adapt_lookup_ci_name() uses strncmp() to perform the lookup, but
applies the length of the current item as maximum length for the string
comparison.
This leads to false acceptance when an item is a prefix of the actually
looked up item, e.g. the item "SOME_REGISTER" matches for the search
string "SOME_REGISTER_BIT4".

The fix looks up the maximum length among *all* knobs, and applies this
common limit to all comparisons. The maximum length is cached in a
function local static variable.