add alderlake versions of performance counters
The existing counters only refer to fixed function counters 0,1,2; but
on Alder Lake a fourth counter (no. 3) is supported. In order to not
break compatibility, an additional, alderlake-exclusive version of the
corresponding knobs is introduced.

For future processors the processor groups for these knobs must be
expanded manually; furthermore renaming the files (currently with suffix
_alderlake) is probably advised.